% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{bigram_counts}
\alias{bigram_counts}
\title{Integer counts of adjacent letter bigrams across a word list}
\usage{
bigram_counts(words)
}
\description{
Integer counts of adjacent letter bigrams across a word list
}
\keyword{internal}
