% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{build_osexp}
\alias{build_osexp}
\title{Build a complete plain-text OpenSesame experiment}
\usage{
build_osexp(design_name, language, conditions_file, schema)
}
\description{
Build a complete plain-text OpenSesame experiment
}
\keyword{internal}
