% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradigms.R
\name{content_field}
\alias{content_field}
\title{If \code{content} is a single braced field reference, return the bare field name}
\usage{
content_field(content)
}
\description{
If \code{content} is a single braced field reference, return the bare field name
}
\keyword{internal}
