% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generate_pseudowords_subsyllabic}
\alias{generate_pseudowords_subsyllabic}
\title{A subsyllabic pseudoword for each base word (with letter-substitution fallback)}
\usage{
generate_pseudowords_subsyllabic(base_words, reference_words)
}
\description{
A subsyllabic pseudoword for each base word (with letter-substitution fallback)
}
\keyword{internal}
