% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{Values; \code{a} is returned unless \code{NULL}.}
}
\value{
\code{a} or \code{b}.
}
\description{
Returns \code{a} unless it is \code{NULL}, in which case it returns \code{b}.
}
\keyword{internal}
