% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{read_config}
\alias{read_config}
\title{Read a YAML configuration file}
\usage{
read_config(path)
}
\arguments{
\item{path}{Path to a YAML file.}
}
\value{
A named list.
}
\description{
Read a YAML configuration file
}
\keyword{internal}
