% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradigms.R
\name{referenced_fields}
\alias{referenced_fields}
\title{The ordered, unique trial fields referenced by an event list's content}
\usage{
referenced_fields(events)
}
\description{
The ordered, unique trial fields referenced by an event list's content
}
\keyword{internal}
