% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querying.R
\name{count_syllables}
\alias{count_syllables}
\title{Orthographic syllable estimate: the number of maximal vowel runs}
\usage{
count_syllables(word)
}
\description{
Orthographic syllable estimate: the number of maximal vowel runs
}
\keyword{internal}
