% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{generate_pseudowords}
\alias{generate_pseudowords}
\title{A length-matched pseudoword for each base word (byte-order processing)}
\usage{
generate_pseudowords(base_words, reference_words)
}
\description{
A length-matched pseudoword for each base word (byte-order processing)
}
\keyword{internal}
