% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{read_csv_utf8}
\alias{read_csv_utf8}
\title{Read a UTF-8 CSV file}
\usage{
read_csv_utf8(path)
}
\arguments{
\item{path}{Path to a CSV file.}
}
\value{
A \link[tibble:tibble]{tibble} data frame.
}
\description{
Read a UTF-8 CSV file
}
\keyword{internal}
